requirements.txt
# --- Ключевые библиотеки ---
diffusers==0.32.2
transformers==4.39.3
accelerate==0.30.1

# --- CUDA-сборка PyTorch (для Python <=3.11) ---
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# --- Остальное ---
Pillow==11.2.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
numpy
tqdm
requests
